Shifted the vector functions to a different file.
The calc.c was getting crowded, so, it made sense to keep the vector
functions in a separate file for future uses. So, behold the vector.c
and the vector.h.
